backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flasgger==0.9.7.1
marshmallow==3.20.1
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
PyJWT==2.10.1
